- Correction d'un gros bug avec la gestion du buffer des agents - Mise à jour de l'UML - Implémentation de la mise à jour de la statistique du taux d'occupation des buffers des agents (à la réception d'un message, à la fin d'un traitement et à la fin de la simulation - Ajout d'une méthode utilitaire

+++ trunk/src/main/java/be/simulation/utilitaires/Utilitaires.java
@@ -24,7 +24,7 @@ public class Utilitaires {
 	 *        le total
 	 * @return le pourcentage
 	 */
-	public static double calculerPourcentage(final int nombre, final int total) {
+	public static double calculerPourcentage(final long nombre, final long total) {
 		long result = 0;
 		if (nombre > 0) {
 			return ((double) nombre / (double) total) * 100;
